backend/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
pymongo==4.6.0
eventlet==0.35.2
PyJWT==2.12.1
Werkzeug==3.0.1
